refactor : LadderGameResult 클래스 레이어 변경  - 서비스 레이어로 이동  - LadderGameService 의 실행결과로 원시타입 String이 아닌 LadderGameResult 인스턴스 반환하도록 수정
