+++ src/main/java/ladder/controller/dto/LadderGenerationResponse.java
@@ -1,5 +1,7 @@
 package ladder.controller.dto;
 
+import ladder.service.dto.LadderGameResult;
+
 import java.util.List;
 
 public class LadderGenerationResponse {
