+++ src/main/java/ladder/service/LadderGameService.java
@@ -1,8 +1,10 @@
 package ladder.service;
 
-import ladder.domain.MatchingItems;
 import ladder.domain.Ladder;
+import ladder.domain.MatchingItems;
+import ladder.domain.Participant;
 import ladder.domain.Participants;
+import ladder.service.dto.LadderGameResult;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -10,11 +12,12 @@ import java.util.List;
 
 public class LadderGameService {
 
-    public List<String> executeGame(Participants participants, Ladder ladder, MatchingItems matchingItems) {
-        List<String> gameResultList = new ArrayList<>();
-        for (int entranceNumber = Participants.MIN_ENTRANCE_NUMBER; entranceNumber < participants.getCount(); entranceNumber++) {
-            gameResultList.add(matchingItems.getGameResult(ladder.getExitNumber(entranceNumber)));
+    public List<LadderGameResult> executeGame(Participants participants, Ladder ladder, MatchingItems matchingItems) {
+        List<LadderGameResult> gameResults = new ArrayList<>();
+        int entranceNumber = Participants.MIN_ENTRANCE_NUMBER;
+        for (Participant participant : participants.getParticipants()) {
+            gameResults.add(new LadderGameResult(participant.getName(), matchingItems.getGameResult(ladder.getExitNumber(entranceNumber++))));
         }
-        return Collections.unmodifiableList(gameResultList);
+        return Collections.unmodifiableList(gameResults);
     }
 }
